% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/A4.measures.R
\name{A4.measures}
\alias{A4.measures}
\title{A4 page measures}
\usage{
A4.measures(metric = c("cm", "inch"))
}
\arguments{
\item{metric}{= c('cm', 'inch')}
}
\description{
A4 page measures
}
\examples{

A4.measures
}
