% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acknowledge.R
\name{acknowledge}
\alias{acknowledge}
\title{Acknowledge
 
Highlight your results with, what else: asterisks!}
\usage{
acknowledge(value, listie)
}
\arguments{
\item{value}{or values. The values you'd like to acknowledge}

\item{list}{a list with the values of interest}
}
\description{
Acknowledge
 
Highlight your results with, what else: asterisks!
}
\examples{
acknowledge
}
