% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alphabet.R
\docType{data}
\name{alphabet}
\alias{alphabet}
\title{Object: a list containing all letters of the alphabet}
\format{An object of class \code{character} of length 26.}
\usage{
alphabet
}
\description{
This object saves you typing 26 characters and 52 quotes.
}
\examples{
alphabet
}
\keyword{datasets}
