% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyse.microbeta.data.R
\name{analyse.microbeta.data}
\alias{analyse.microbeta.data}
\title{Analyse your micro.beta data.}
\usage{
analyse.microbeta.data(data, pool.plates = T, screen.type = "hit",
  pos.control = "ATP", baseline = "BSA", norm.over.ATP = T,
  set.to.zero = F, nr.plates = 5, plate.type = 96,
  graph.levelplots = F, path = path,
  levelplot.file.names = "Default", estimate.p.values = F,
  plot.levelplots = F, p.value.distribution = c("chisq", "gamma",
  "exp", "normal"), alpha = 0.05, plot.distr.estimate = F,
  graph.density = F)
}
\arguments{
\item{data}{Data that were loaded using load.microbeta and subsequently summarised using process.microbeta.data()}

\item{pool.plates}{would you like to pool the plates in your assay? IF not, you'll get one plot as output, otherwise it produces a list of several plots.}

\item{screen.type}{Default= wide screen.}

\item{pos.control}{= "ATP", your positive control.}

\item{baseline}{= "BSA", the solution in which peptides were diluted.}

\item{norm.over.ATP}{Default = F: Would you like to divide your values by the ATP value per plate or not?}

\item{set.to.zero}{Default = T: would you like to set negative z.scores to 0?}

\item{nr.plates}{The number of plates you screened.}

\item{plate.type}{The type of plates you used (i.e. e.g. 96 wells)}

\item{graph.levelplots}{Default: F; If T: writes a levelplot graph to the specified location in path}

\item{path}{If you opted to write graphs to ppt: where should the file go?}

\item{levelplot.file.names}{Default: 'Levelplot Plate 1', 2, 3, ... If graph2ppt: how should you like to call your files?}

\item{estimate.p.values}{Default = F: Would you like to estimate p.values?}

\item{plot.levelplots}{If T: levelplots are plotted, this can take a while}

\item{p.value.distribution}{No Default: what distribution do you think you'r data has?}

\item{alpha}{What alpha value would you like to identify hits on?}

\item{plot.distr.estimate}{Default = F; Would you like to plot the estimated distribution of your data?}

\item{graph.density}{Default = F; Woudl you like to write the density curve to a file?}
}
\description{
Using this function, you can elegantly and rather quickly scout your microbeta screen for potential hits. The outcome will be a list of several elements, including a density plot, the potential hits and a levelplot.
}
\examples{

analyse.microbeta.data
}
