% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyse.MSA.R
\name{analyse.MSA}
\alias{analyse.MSA}
\title{Analyse MSA data}
\usage{
analyse.MSA(folder = getwd(), destination = getwd(), writeCsv = T)
}
\arguments{
\item{folder}{folder in which you saved the python output}

\item{destination}{folder in which you'd like to save the output files}

\item{writeCsv}{if you want to save the output files in a csv.}
}
\description{
Analyse MSA data which you first fed into the python script "analyse_alignment_on_similarity_in_speciesv2.py"
}
\examples{
analyse.MSA
}
