% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asterisks.R
\name{asterisks}
\alias{asterisks}
\title{Symbolizing significance by asterisks}
\usage{
asterisks(value, cutpoints = c(0, 0.001, 0.01, 0.05, 0.1, 1),
  symbols = c("***", "**", "*", ".", ""))
}
\arguments{
\item{value}{Here goes your data.}

\item{cutpoints}{Here goes a vector of cutpoints, usually one longer than the vector of thresholds}

\item{symbols}{Here goes a vector of symbols, one shorter than the vector of cutpoints}
}
\description{
Adapted from prof. Tom Wenseleers:
This function uses given threshold values (cutpoints) to code these in symbols (symbols)
By default, it will annote asterisks (*) to p.value values.
}
\examples{
asterisks
}
