% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binning.R
\name{binning}
\alias{binning}
\title{Replace all NA values by a constant.}
\usage{
binning(data, binrange, set_bins = "automatic", manual_bins = c())
}
\arguments{
\item{data}{Here you put the data for which you like to replace it by bins.}

\item{binrange}{Here you put the range that the bins should span. E.g. if you want to put bins for years, you can chose 5 years and that'll give you bins of five years wide.}

\item{set_bins}{Optinoal argument: 'automatic', or 'manual'; Default: 'automatic'. If you like to set bins manually, select manual.}

\item{manual_bins}{Here goes a vector with the bins you'd like. If you already selected automatic, this argument is optional.}
}
\description{
This function enables you to replace all NA values by a constant.
}
\examples{
binning
}
