% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.folder.R
\name{check.folder}
\alias{check.folder}
\title{Show all files in a folder}
\usage{
check.folder(folder = getwd())
}
\arguments{
\item{folder}{Default = wd. What folder do you want to check the files of?}
}
\description{
Check what files are in a folder.
}
\examples{
check.folder
}
