% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.learningData.R
\name{check.learningData}
\alias{check.learningData}
\title{Checking your learning data whether it worked or not.}
\usage{
check.learningData(data, column, strain = "N2", learning.phenotype = T,
  threshold = 0.7, learning.cutpoints = c(-1, (threshold - 0.2),
  (threshold - 0.1), (threshold - 0.05), (threshold), 1),
  non.learning.cutpoints = c(-1, (threshold - 0.15), (threshold),
  (threshold + 0.1), (threshold + 0.2), 1))
}
\arguments{
\item{data}{Here goes your data, loaded by load.learningdata()}

\item{column}{data$column: Here goes the column you'd like to check: data$CI.}

\item{strain}{what strain would you like to check up on? Default: N2}

\item{learning.phenotype}{Is the strain learning (true) or shouldn't it be (false)? Default: True (learning)}

\item{threshold}{The threshold CI value for which the assay went well. For naive: CI > threshold is good, whereas for conditioned CI < threshold is better}

\item{learning.cutpoints}{The cutpoints you used to determine the significance of the learning phenotype}

\item{nonlearning.cutpoints}{The cutpoints you used to determine the significance of the non-learning phenotype}
}
\description{
This function creates a data.frame in which you can check the quality of your learning assay.
}
\examples{
check.learningData
}
