% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chemoindex.R
\name{chemoindex}
\alias{chemoindex}
\title{Calculcating the Chemotaxis index}
\usage{
chemoindex(DA, EtOH, Others)
}
\arguments{
\item{DA}{is the number of worms in the diacetyl spot}

\item{EtOH}{is the number of worms in the ethanol spot}

\item{Others}{is the remainder of worms on the plate}
}
\description{
This function calculates a chemotaxis index
}
\examples{
load.LocomotionData
}
