% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cm2inch.R
\name{cm2inch}
\alias{cm2inch}
\title{Convert cm to inches}
\usage{
cm2inch(value)
}
\arguments{
\item{value}{The value you'd like to convert to inches}
}
\description{
Convert cm to inches. What else to explain?
}
\examples{
cm2inch
}
