% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cumulate.R
\name{cumulate}
\alias{cumulate}
\title{Cumulate
 
This function cumulates ordered stuff}
\usage{
cumulate(values)
}
\arguments{
\item{values}{The values you'd like to cumulate}
}
\description{
Cumulate
 
This function cumulates ordered stuff
}
\examples{
cumulate
}
