% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/document.WormTools.R
\name{document.WormTools}
\alias{document.WormTools}
\title{Document and install your newly written code in one line!}
\usage{
document.WormTools()
}
\description{
This simple function documents your newly written function in one line to the WormTools directory.
}
\examples{
document.WormTools
}
