% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/folderTag.R
\name{folderTag}
\alias{folderTag}
\title{Produce a tag for every subfolder in a folder}
\usage{
folderTag(folder = getwd(), suffix = "")
}
\arguments{
\item{folder}{This can be any folder. By default the function will check all subfolers the wd}
}
\description{
As the title states: this function will go over all subfolders in a folder and produce an identifier for each folder
}
\examples{
folderTag
}
