% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/invert.dataframe.R
\name{invert.dataframe}
\alias{invert.dataframe}
\title{Invert your data frame
 
This function inverts data.frames}
\usage{
invert.dataframe(dataframe)
}
\arguments{
\item{dataframe}{a data.frame which you'd like to see inverted.}
}
\description{
Invert your data frame
 
This function inverts data.frames
}
\examples{
invert.dataframe
}
