% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/levelplot.data.microbeta.R
\name{levelplot.data.microbeta}
\alias{levelplot.data.microbeta}
\title{Produce data for a levelplot for your receptor screen data}
\usage{
levelplot.data.microbeta(data, nr.plates = 5, plate.type = 96,
  nr.x = 12, nr.y = 8)
}
\arguments{
\item{data}{This is the data you loaded with load.microbeta() and processed with process.microbeta.data()}

\item{pos.control}{Default: 'ATP', this is the character string coding for the positive control (I wasn't entirely sure whether this is always ATP)}

\item{baseline}{Default: 'BSA'. Idem, but for teh baseline values, i.e. the solvent in which you diluted your peptides.}

\item{norm.over.ATP}{Default: True, this is an optional argument that you can tick if you'd like to normalize over the positive control.}
}
\description{
Returns a data.frame for receptor data to produce a levelplot after you processed them with process.microbeta.data
}
\examples{
levelplot.data.microbeta
}
