% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/levelplot.microbeta.R
\name{levelplot.microbeta}
\alias{levelplot.microbeta}
\title{Make a levelplot (i.e. a heatmap without clustering) for your microbeta data}
\usage{
levelplot.microbeta(data, pool.plates = F, nr.plates = 5,
  plate.type = 96, graph2ppt = F, graph.names = "Default",
  path = getwd())
}
\arguments{
\item{data}{Data that were loaded using load.microbeta and subsequently summarised using process.microbeta.data()}

\item{pool.plates}{would you like to pool the plates in your assay? IF not, you'll get one plot as output, otherwise it produces a list of several plots.}

\item{nr.plates}{The number of plates you screened.}

\item{plate.type}{The type of plates you used (i.e. e.g. 96 wells)}

\item{graph2ppt}{Default: F; If T: writes a graph to the specified location in path}

\item{path}{If graph2ppt: where should the file go?}

\item{file.names}{Default: 'Levelplot Plate 1', 2, 3, ... If graph2ppt: how should you like to call your files?}
}
\description{
This function will allow you to plot your microbeta data to visually see whether you have potential hits.
}
\examples{
levelplot.microbeta
}
