% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.layout.R
\name{load.layout}
\alias{load.layout}
\title{Loading the plate layout.}
\usage{
load.layout(filename, nr.plates = 5)
}
\arguments{
\item{filename}{The name of the file.}

\item{nr.plates}{Default == 5: the number of plates you screened.}
}
\description{
This function loads the plate layout and adds the relevant columns to the dataset. Load your .csv file as a data.frame with a first column in which the y-values (A, B, ..., H) are and the next few columns show the peptide per x-value (1, 2, ..., 12). All plates are pasted underneath eachother.
}
\examples{
load.layout
}
