% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.LocomotionData.R
\name{load.LocomotionData}
\alias{load.LocomotionData}
\title{Loading data from the Locomotion Assay}
\usage{
load.LocomotionData(path = getwd(), folders, strains, conditions)
}
\arguments{
\item{path}{Enter the path to the main folder.}

\item{folders}{Enter all subfolders you constituted. I'll try to find a way to do this more efficiently.}

\item{strains}{Enter a list of the strains you used. These should be ordered in the way you ordered the respective folders.}

\item{conditions}{Enter a list of again respective conditions for all folders. Mind that the order does matter a lot!}
}
\description{
This function allows you to quickly load all data from one mother folder in which you saved data for several experiments in subfolders. E.g. the main folder contains your data, the subfolders contain .csv files for each time a different strain in a different condition.
}
\examples{
load.LocomotionData
}
