% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.microbeta.R
\name{load.microbeta}
\alias{load.microbeta}
\title{Loading the data from the microbeta machine.}
\usage{
load.microbeta(data, layout, nr.plates = 5, nr.series = 30,
  matching = "automatic", filename = "microbeta.compiled.data.csv",
  write.csv = T)
}
\arguments{
\item{data}{The data from your receptor screen}

\item{layout}{Optional argument: The plate layout.}

\item{nr.plates}{The number of plates you screened.}

\item{nr.series}{The number of seconds in the peptide addition and triton addition measurements. Default: 30 (on microbeta)}

\item{matching}{== c('manual', 'automatic') Default: 'automatic' Optional argument  that by default loads in the plate layout as well. THIS CAN TAKE A WHILE.}

\item{filename}{The name you'd like to give to the .csv file that you'll write.}

\item{write.csv}{== c(T, F), Default: T. Optional argument to write a .csv file with your data.}
}
\description{
This function loads your .csv file from the microbeta machine with an option to immediately read in the peptides from the plate layout file.
}
\examples{
load.microbeta
}
