% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.MultipleCsv.R
\name{load.MultipleCsv}
\alias{load.MultipleCsv}
\title{Loading multiple .csv files from a common folder.}
\usage{
load.MultipleCsv(add.replicate = F)
}
\arguments{
\item{NONE}{Be sure to set your working directory to the folder of which you want to access the data. Your data should be formatted as follows: One column in which there is the data for the DA spot, one column in which there is the data for the EtOH spot and another column in which there is the data for all other worms. Next to this, the function requires a column that indicates the strain and a column that indicates the condition.}
}
\description{
This function allows you to quickly load all data from one mother folder in which you saved data as .csv files.
}
\examples{
load.MultipleCsv
}
