% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make.a.range.R
\name{make.a.range}
\alias{make.a.range}
\title{Make a range}
\usage{
make.a.range(values)
}
\arguments{
\item{values}{the vector of values in which you see several ranges}
}
\description{
Make a range out of a vector of elements. This allows you to visually make it easier to find gropus of subsequent numbers.
}
\examples{
make.a.range
}
