% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/na.col.omit.R
\name{na.col.omit}
\alias{na.col.omit}
\title{Omitting NA values (or rows) for a specific column of your dataset}
\usage{
na.col.omit(data, column)
}
\arguments{
\item{data}{Here goes your data. Accepted datatypes e.g. data.frame, lists, ...}

\item{column}{data$column: Here goes the column you'd like to check.}
}
\description{
Removing NA values from a specific column of your dataset: If it's a data.frame, the row is omitted, if it is a
}
\examples{
NaN.omit
}
