% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/na.replace.R
\name{na.replace}
\alias{na.replace}
\title{Replacing NA values by a constant.}
\usage{
na.replace(data, replacement)
}
\arguments{
\item{data}{Here goes the data}

\item{replacement}{Here goes your replacement}
}
\description{
This function allows you to replace any NA that occur by a value. Chose the value wisely so that it will not interfere with your data, but will allow you to e.g. divide it or execute logical statements.
}
\examples{
na.replace
}
