% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.blank.R
\name{plot.blank}
\alias{plot.blank}
\title{Plot a blank screen}
\usage{
plot.blank()
}
\description{
Sometimes the plotting window does a little onorthodox things and you might want to plot a blank screen in between.
}
\examples{
plot.blank
}
