% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.microbeta.data.R
\name{process.microbeta.data}
\alias{process.microbeta.data}
\title{Check for hits in your screening data.}
\usage{
process.microbeta.data(data, pos.control = "ATP", baseline = "BSA",
  norm.over.ATP = T, pool.plates = T, return.list = F,
  set.to.zero = F)
}
\arguments{
\item{data}{This is the data you loaded with load.microbeta()}

\item{pos.control}{Default: 'ATP', this is the character string coding for the positive control (I wasn't entirely sure whether this is always ATP)}

\item{baseline}{Default: 'BSA'. Idem, but for teh baseline values, i.e. the solvent in which you diluted your peptides.}

\item{norm.over.ATP}{Default: True, this is an optional argument that you can tick if you'd like to normalize over the positive control.}

\item{pool.plates}{Default: True, this is an optional argument that allows you to run the analysis on each plate separately if you'd like to.}

\item{return.list}{Default: False, this is an optional argument that allows you to return a list if you chose to run the analysis on separate plates. If not, remember to keep plates separate throughout your graphs.}

\item{set.to.zero}{Default: False, this is an optional argument which you can select if you'd like to put negative values to zero.}
}
\description{
Returns a data.frame to identify hits: If you ran and loaded your peptide data with load.microbeta(), this function allows you to screen for hits. For now, hits are defined when the z-score is 2 standarddeviations (*) away from the mean of the population of hits for the peptide data only (excluding ATP and baseline BSA levels). However, for now, I haven't found a distribution that fits this plot.
}
\examples{
process.microbeta.data
}
