% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set.to.zero.R
\name{set.to.zero}
\alias{set.to.zero}
\title{Set a series of values to 0 (or a threshold value)}
\usage{
set.to.zero(value, output = 0, threshold = 0)
}
\arguments{
\item{value}{The values you want to set to a threshold if}

\item{output}{Choose the output value you'd like.}

\item{threshold}{Choose the threshold}
}
\description{
This function will set any value to an arbitrary value if below a threshold value (for now). If you enter a column in a dataset, mind to indeed enter the column!
}
\examples{
set.to.zero
}
