% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setwd.WormTools.R
\name{setwd.WormTools}
\alias{setwd.WormTools}
\title{Set your working directory to the WormTools root directory.}
\usage{
setwd.WormTools()
}
\description{
This simple function just sets your working directory to the WormTools root directory. Therefore, it will only work if you switched to administrator mode.
}
\examples{
setwd.WormTools
}
