% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/string.find.R
\name{string.find}
\alias{string.find}
\title{find a substring in a string}
\usage{
string.find(string, substring, mode = "first")
}
\arguments{
\item{string}{The string you want to find it in}

\item{substring}{The substring you want to find}

\item{mode}{= c('first', 'all'): only the first occurence or all occurences?}
}
\description{
Similar to the str.find command in Python: find the index of a substring in a character string.
}
\examples{
string.find
}
