% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/string.substringLeft.R
\name{string.substringLeft}
\alias{string.substringLeft}
\title{Generate a head of a string!}
\usage{
string.substringLeft(x, n)
}
\arguments{
\item{x}{A string you'd like to slice}

\item{n}{the number of characters your string should contain.}
}
\description{
Generate a substring composed of the first n characters of your string in your string of interest
}
\examples{
string.substringLeft
}
