% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/string.substringRight.R
\name{string.substringRight}
\alias{string.substringRight}
\title{Generate a tail of a string!}
\usage{
string.substringRight(x, n)
}
\arguments{
\item{x}{A string you'd like to slice}

\item{n}{the number of characters your string should contain.}
}
\description{
Generate a substring composed of the final n characters of your string in your string of interest
}
\examples{
string.substringRight
}
