% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update.all.packages.R
\name{update.packages}
\alias{update.packages}
\title{This function will update/install a concise list of packages in R}
\usage{
update.packages()
}
\description{
This function will update/install a concise list of packages in R
}
\examples{
update.packages()
}
