% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update.WormTools.R
\name{update.WormTools}
\alias{update.WormTools}
\title{Update the WormTools package}
\usage{
update.WormTools()
}
\description{
This simple function updates the WormTools package in one line
}
\examples{
update.WormTools
}
