% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.txt.R
\name{write.txt}
\alias{write.txt}
\title{Write output to a .txt file}
\usage{
write.txt(data, file, suffix = ".txt", folder = getwd())
}
\arguments{
\item{data}{The data you'd like to output. Make sure this is a character string.}

\item{file}{The name you want your file to carry}

\item{suffix}{By default: .txt .If you'd like to rather safe it as another file than .txt, take in mind the spacer then.}

\item{folder}{The destination folder, by default: working directory}
}
\description{
Write your output to a .txt file
}
\examples{
write.txt
}
