% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z.score.R
\name{z.score}
\alias{z.score}
\title{Calculate the z-score.}
\usage{
z.score(value, mean, sd, var, method = "sd")
}
\arguments{
\item{value}{The value of which you'd like to calculate the z-score.}

\item{mean}{The mean value}

\item{sd}{The sd of the distribution of your value.}

\item{var}{The variance of the distribution. By default, the sd is used, but if you like you can also put the variance in;}

\item{method=='sd'}{By default you put the sd in, however, the variance can also work, if you indicate that you're using the variance;}
}
\description{
Calculate the z-score of a value by manually indicating the mean and the sd/var.
}
\examples{
z.score
}
