% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ZI.match.R
\name{ZI.match}
\alias{ZI.match}
\title{Matching two datafiles to eachother.}
\usage{
ZI.match(data1, data2, value, type = "numeric")
}
\arguments{
\item{data1}{A vector or data.frame column in which the identifier of data.frame 1 is.}

\item{data2}{A vector or data.frame column in which the identifier of data.frame 2 is.}

\item{value}{A vector or data.frame column in which you can find the values you'd like to return in an output vector.}

\item{type}{Default: numeric, other options: 'character', make sure to select the option character if it is a character.}
}
\description{
If you have two datafiles with overlapping information, this function allows you to select the columns in which the identifier resides for both dataframes and produces a vector in which you put the values of either of the data.frame which you'd like to convert to the other data.frame. Therefore, this method also works if the nrow() differs for both data.frames.
}
\examples{
ZI.match
}
