% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ZI.plot.R
\name{ZI.plot}
\alias{ZI.plot}
\title{Plot in Nathan's favourite style (hint: There are options!).}
\usage{
ZI.plot(data, x, y, group, fill, colour = fill, facet_wrap = F,
  facet = "", ylab = "y", xlab = "x", filling = "automatic",
  colours, type = "Boxplot")
}
\arguments{
\item{data}{Here goes your data as a data-frame.}

\item{x}{x-values (format: either data$column or column)}

\item{y}{y-values (format: idem)}

\item{group}{How you'd like to group your variables.}

\item{fill}{How you'd like to fill the plots. Different from colour.}

\item{colour}{Default == fill. How you'd like to colour the outlines of your plotted elements (e.g. outline of boxes in the boxplot)}

\item{facet_wrap}{(T or F) Would you like to have several panels?}

\item{facet}{Optional: If you chose for several panels, here you indicate what the panels should include.}

\item{ylab}{Labels the y-axis}

\item{xlab}{Labels the x-axis}

\item{filling}{'manual' or 'automatic' (Default: automatic) How would you like to fill the plots? Manually or automatically?}

\item{colours}{If you opted filling = 'manual', provide a vector of the colours you'd like. This should have the same length as the number of objects you graph.}

\item{type}{'Boxplot' or 'violin' for now.}

\item{colour}{How you'd like to colour your}
}
\description{
This function elegantly puts together a graph using ggplot. 
Be sure ggplot2 and ggthemes are installed.
}
\examples{
ZI.plot
}
