% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ZI.posthoc.R
\name{ZI.posthoc}
\alias{ZI.posthoc}
\title{Produce a post-hoc analysed data.frame for your fit.}
\usage{
ZI.posthoc(fit, factors, nr.factors = 2, method = "pairwise",
  adjust = "tukey", selected.contrasts = F, contrasts)
}
\arguments{
\item{fit}{The fit for which you'd like to produce a posthoc analysis. Don't forget to first run an Anova on it.}

\item{factors}{The factors you'd like to check contrasts for. Give a vector containing all factors.}

\item{nr.factors}{= 2 By default: nr.factor == 2. Chose c(1, 2, 3, 4)}

\item{method}{= "pairwise" by default. Others are possible (check ?contrast)}

\item{adjust}{= "tukey" by default. An adjustment of your p-values is usually required.}
}
\description{
Produce a post-hoc analysed data.frame for your fit.
}
\examples{
ZI.posthoc
}
